Added tests for RHAI Manage page
